Addition of overload argument to the SelectElement.SelectByText method
Issue raised in the issue log label as dotnet, issue number 3575. The
XML summary informs that the method provides partial match ability on
the options  by Text.  However this wasn't fully implemented so pushed
the ability via a boolean to perform such a partial match on an options
list.

Adding in the App.config change in order to run tests via the IDE for
the Support project.

Fixes #3575