Addition of Test Playlists for Each Test Project in the Repo.
This is to help aid run all tests in a specific project. I was commiting a fix
for a Support UI dotnet issue and spoke with Jim about the addition of playlists.
This will help future people in the dotnet bindings easily identify and play all tests in a
playlist.

Fixes #3805